Fix sorting tags test case.
This test failed as the meta value in the hashes were arrays with
elements in different orders. I'm not sure if this matters but the test
description makes suggests that perhaps order is important here and this
is in fact a real failure. I've left a comment in order to do more
investigation at a future point.